users-authentication 3-7 be kar giri takhsis mojavez ba gate ha

+++ app/Http/Controllers/PostController.php
@@ -6,6 +6,7 @@ use App\Like;
 use App\Post;
 use App\Tag;
 use Auth;
+use Gate;
 use Illuminate\Http\Request;
 
 class PostController extends Controller
@@ -89,6 +90,9 @@ class PostController extends Controller
             'content' => 'required|min:10'
         ]);
         $post = Post::find($request->input('id'));
+        if (Gate::denies('manipulate-post', $post)) {
+            return redirect()->back();
+        }
         $post->title = $request->input('title');
         $post->content = $request->input('content');
         $post->save();
@@ -104,6 +108,9 @@ class PostController extends Controller
             return redirect()->back();
         }
         $post = Post::find($id);
+        if (Gate::denies('manipulate-post', $post)) {
+            return redirect()->back();
+        }
         $post->likes()->delete();
         $post->tags()->detach();
         $post->delete();
